// src/app/types/flight.ts
interface Flight {
    flight_number?: number;
    mission_name?: string;
    launch_year?: string;
    details?: string;
    rocket_name?: string;
    rocket_type?: string;
    mission_patch?: string;
    article_link?: string;
    wikipedia?: string;
    video_link?: string;
}

export default Flight